requirements.txt
fastapi
uvicorn[standard]
sqlmodel
jinja2
python-multipart
requests
pyyaml
python-nmap